% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augADMMExample.R
\name{augADMMExample}
\alias{augADMMExample}
\title{Add Two Numbers}
\usage{
augADMMExample(num1, num2)
}
\arguments{
\item{num1}{A numeric value.}

\item{num2}{A numeric value.}
}
\value{
The sum of num1 and num2.
}
\description{
This function takes two numeric values and returns their sum.
}
\examples{
augADMMExample(5, 3) # returns 8

}
